+++ packages/core-js-compat/src/data.js
@@ -854,7 +854,7 @@ const data = {
   },
   'es.string.match-all': {
     // Early implementations does not throw an error on non-global regex
-    // chrome: '73',
+    chrome: '80', // 73
     // firefox: '67',
     // safari: '13',
   },