+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ##### Unreleased
 - Added a workaround of V8 deoptimization which causes serious performance degradation (~20x in my tests) of some `RegExp`-related methods like `String#split`, [#306](https://github.com/zloirock/core-js/issues/306)
 - Fixed work of `fetch` with polyfilled `URLSearchParams`, [#674](https://github.com/zloirock/core-js/issues/674)
+- Added compat data for Chrome 80
 - `package-lock.json` no longer generated in libraries
 
 ##### 3.3.3 - 2019.10.22